requirements.txt
prophet==1.1.5
pandas==2.1.4
mysql-connector-python==8.3.0
PyMySQL==1.1.0
SQLAlchemy==2.0.23
pyOpenSSL==24.0.0
